+++ packages/iov-bns-governance/src/governor.spec.ts
@@ -42,7 +42,7 @@ async function getGovernorOptions(
   path = adminPath,
 ): Promise<GovernorOptions & { readonly profile: UserProfile }> {
   const connection = await BnsConnection.establish(bnsdUrl);
-  const chainId = await connection.chainId();
+  const chainId = connection.chainId;
   const profile = new UserProfile();
   const wallet = profile.addWallet(Ed25519HdWallet.fromMnemonic(adminMnemonic));
   const identity = await profile.createIdentity(wallet.id, chainId, path);