+++ packages/iov-cosmos/src/cosmosconnection.ts
@@ -41,10 +41,6 @@ import { RestClient, TxsResponse } from "./restclient";
 
 const { fromBase64 } = Encoding;
 
-interface ChainData {
-  readonly chainId: ChainId;
-}
-
 // poll every 0.5 seconds (block time 1s)
 const defaultPollInterval = 500;
 
@@ -72,17 +68,13 @@ function buildQueryString({
 export class CosmosConnection implements BlockchainConnection {
   public static async establish(url: string): Promise<CosmosConnection> {
     const restClient = new RestClient(url);
-    const chainData = await this.initialize(restClient);
-    return new CosmosConnection(restClient, chainData);
-  }
-
-  private static async initialize(restClient: RestClient): Promise<ChainData> {
     const { node_info } = await restClient.nodeInfo();
-    return { chainId: Caip5.encode(node_info.network) };
+    const chainId = Caip5.encode(node_info.network);
+    return new CosmosConnection(restClient, chainId);
   }
 
+  public readonly chainId: ChainId;
   private readonly restClient: RestClient;
-  private readonly chainData: ChainData;
   private readonly primaryToken: Token;
   private readonly supportedTokens: readonly Token[];
 
@@ -90,9 +82,9 @@ export class CosmosConnection implements BlockchainConnection {
     return "cosmos";
   }
 
-  private constructor(restClient: RestClient, chainData: ChainData) {
+  private constructor(restClient: RestClient, chainId: ChainId) {
+    this.chainId = chainId;
     this.restClient = restClient;
-    this.chainData = chainData;
     this.primaryToken = {
       fractionalDigits: 6,
       tokenName: "Atom",
@@ -103,10 +95,6 @@ export class CosmosConnection implements BlockchainConnection {
 
   public disconnect(): void {
     return;
-  }
-
-  public chainId(): ChainId {
-    return this.chainData.chainId;
   }
 
   public async height(): Promise<number> {
@@ -182,8 +170,7 @@ export class CosmosConnection implements BlockchainConnection {
   ): Promise<ConfirmedAndSignedTransaction<UnsignedTransaction> | FailedTransaction> {
     try {
       const response = await this.restClient.txsById(id);
-      const chainId = await this.chainId();
-      return this.parseAndPopulateTxResponse(response, chainId);
+      return this.parseAndPopulateTxResponse(response, this.chainId);
     } catch (error) {
       if (error.response.status === 404) {
         throw new Error("Transaction does not exist");
@@ -235,9 +222,8 @@ export class CosmosConnection implements BlockchainConnection {
     query: TransactionQuery,
   ): Promise<readonly (ConfirmedTransaction<UnsignedTransaction> | FailedTransaction)[]> {
     const queryString = buildQueryString(query);
-    const chainId = this.chainId();
     const { txs: responses } = await this.restClient.txs(queryString);
-    return Promise.all(responses.map(response => this.parseAndPopulateTxResponse(response, chainId)));
+    return Promise.all(responses.map(response => this.parseAndPopulateTxResponse(response, this.chainId)));
   }
 
   public listenTx(