+++ packages/iov-lisk/types/liskconnection.d.ts
@@ -26,11 +26,10 @@ import { Stream } from "xstream";
 export declare function generateNonce(): Nonce;
 export declare class LiskConnection implements BlockchainConnection {
   static establish(baseUrl: string): Promise<LiskConnection>;
+  readonly chainId: ChainId;
   private readonly baseUrl;
-  private readonly myChainId;
   constructor(baseUrl: string, chainId: ChainId);
   disconnect(): void;
-  chainId(): ChainId;
   height(): Promise<number>;
   postTx(bytes: PostableBytes): Promise<PostTxResponse>;
   getToken(searchTicker: TokenTicker): Promise<Token | undefined>;