+++ CHANGELOG.md
@@ -41,6 +41,8 @@ Breaking changes
   `creator`.
 - @iov/bcp: Replace `primarySignature` and `otherSignatures` in
   `SignedTransaction` with a single `signatures` field of type `NonEmptyArray`.
+- @iov/bcp: Convert getter function `BlockchainConnection.chainId()` into
+  read-only property `.chainId`.
 - @iov/bns: Require fee payer to be specified when encoding a transaction with a
   fee.
 - @iov/bns: Use single `signatures` field in `SignedTransaction`s.